Reorganized kernel drivers by domain (bus/graphics/input/interrupts/network/platform/storage/usb) and updated include paths/Makefiles accordingly. Moved core input modules under kernel/source/input and kernel/include/input (Key, MouseCommon, MouseDispatcher, VKey). Moved InterruptController under drivers/interrupts and adjusted dependent modules. Added HID consumer control support for keyboard media keys (usage page 0x0C) in USB keyboard path. Added reusable HID report descriptor parser helper (utils/HIDReport). Extended virtual key set with media keys and added key up/down routing for virtual-key driven events. Replaced system mouse-info app with input-info and updated boot image packaging. Reduced/fixed compiler warnings across kernel/runtime, including x86-64 specific warnings. Fixed dashboard VESA/dumper test flow and added scripts/run.sh --no-nvme option. Minor boot/debug guard cleanup (DEBUG_OUTPUT wrapping in boot multiboot path).

+++ kernel/source/MemoryEditor.c
@@ -25,9 +25,9 @@
 #include "Base.h"
 #include "Console.h"
 #include "Kernel.h"
-#include "drivers/Keyboard.h"
+#include "drivers/input/Keyboard.h"
 #include "CoreString.h"
-#include "VKey.h"
+#include "input/VKey.h"
 
 /***************************************************************************/
 
@@ -42,7 +42,7 @@ static void PrintMemoryLine(U32 Base) {
     STR Addr[16];
     STR Hexa[64];
     STR ASCII[64];
-    U8* Pointer = (U8*)Base;
+    U8* Pointer = (U8*)(UINT)Base;
     U8 Data = 0;
     U32 Index = 0;
